requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
streamlit==1.28.0
requests==2.31.0
tensorflow>=2.10.0
opencv-python==4.8.1.78
numpy>=1.21.0
Pillow>=9.0.0
imageio==2.31.1

